requirements.txt
django
djangorestframework
drf-yasg
django-rest-swagger